marked point series & gpl meta
